chore(client): fix file miscommunication. add triplistwidgets
