+++ src/client/components/index.js
@@ -52,6 +52,8 @@ export StarRatingWidget from './StarRatingWidget/StarRatingWidget'
 export SendButtonIntl from './SendButtonIntl/SendButtonIntl'
 export TextIntroPlainWidget from './TextIntroPlainWidget/TextIntroPlainWidget'
 export TextIntroRichWidget from './TextIntroRichWidget/TextIntroRichWidget'
+export TripListWidget from './TripListWidget/TripListWidget'
+export TripListItemWidget from './TripListItemWidget/TripListItemWidget'
 export TrippianListWidget from './TrippianListWidget/TrippianListWidget'
 export TrippianListItemWidget from './TrippianListItemWidget/TrippianListItemWidget'
 export TrippianListRoundWidget from './TrippianListRoundWidget/TrippianListRoundWidget'
